requirements.txt
google-genai>=1.0.0
faster-whisper>=1.0.0
piper-tts>=1.4.0
sounddevice>=0.4.0
numpy>=1.24.0
